增加按钮控制

+++ trace/src/main/java/com/dili/trace/service/impl/ApproverInfoServiceImpl.java
@@ -2,6 +2,8 @@ package com.dili.trace.service.impl;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +20,7 @@ import com.dili.trace.service.Base64SignatureService;
 
 @Service
 public class ApproverInfoServiceImpl extends BaseServiceImpl<ApproverInfo, Long> implements ApproverInfoService {
-	private static final int MAX_LENGTH=1000;
+	private static final int MAX_LENGTH = 1000;
 
 	@Autowired
 	Base64SignatureService base64SignatureService;
@@ -31,23 +33,18 @@ public class ApproverInfoServiceImpl extends BaseServiceImpl<ApproverInfo, Long>
 		if (StringUtils.isBlank(signBase64)) {
 			throw new BusinessException("签名不能为空");
 		}
+		boolean existSameUserName=this.findByUserName(input.getUserName()).map((item)->{return true;}).orElse(false);
+		if(existSameUserName) {
+			throw new BusinessException("用户名已存在");
+		}
 		this.insertSelective(input);
 
+		List<Base64Signature> base64SignatureList = this.buildBase64SignatureList(signBase64, input.getId());
 
-		List<Base64Signature> base64SignatureList = new ArrayList<>();
-
-		for (int start = 0; start < signBase64.length(); start += MAX_LENGTH) {
-			String base64 = signBase64.substring(start, Math.min(signBase64.length(), start + MAX_LENGTH));
-			Base64Signature base64Signature = DTOUtils.newDTO(Base64Signature.class);
-			base64Signature.setBase64(base64);
-			base64Signature.setApproverInfoId(input.getId());
-			base64Signature.setOrderNum(base64SignatureList.size());
-			base64SignatureList.add(base64Signature);
-		}
-
-		base64SignatureService.batchInsert(base64SignatureList);
+		this.base64SignatureService.batchInsert(base64SignatureList);
 		return 1;
 	}
+
 	@Transactional
 	@Override
 	public int updateApproverInfo(ApproverInfo input) {
@@ -56,37 +53,53 @@ public class ApproverInfoServiceImpl extends BaseServiceImpl<ApproverInfo, Long>
 		if (StringUtils.isBlank(signBase64)) {
 			throw new BusinessException("签名不能为空");
 		}
-		ApproverInfo approverInfo=this.get(input.getId());
+		ApproverInfo approverInfo = this.get(input.getId());
+		boolean existSameUserName=this.findByUserName(input.getUserName()).map((item)->{return (item.getUserName().equals(input.getUserName())&&!item.getId().equals(input.getId()));}).orElse(false);
+		if(existSameUserName) {
+			throw new BusinessException("用户名已存在");
+		}
+		
 		approverInfo.setUserName(input.getUserName());
 		approverInfo.setUserId(input.getUserId());
 		approverInfo.setPhone(input.getPhone());
-		
+
 		this.updateSelective(approverInfo);
-		String oldSignBase64=this.base64SignatureService.findBase64SignatureByApproverInfoId(approverInfo.getId());
-		
-		if(!signBase64.equals(oldSignBase64)) {
-			Base64Signature condition=DTOUtils.newDTO(Base64Signature.class);
+		String oldSignBase64 = this.base64SignatureService.findBase64SignatureByApproverInfoId(approverInfo.getId());
+
+		if (!signBase64.equals(oldSignBase64)) {
+			Base64Signature condition = DTOUtils.newDTO(Base64Signature.class);
 			condition.setApproverInfoId(approverInfo.getId());
 			this.base64SignatureService.deleteByExample(condition);
 
-			List<Base64Signature> base64SignatureList = new ArrayList<>();
+			List<Base64Signature> base64SignatureList = this.buildBase64SignatureList(signBase64, approverInfo.getId());
 
-			for (int start = 0; start < signBase64.length(); start += MAX_LENGTH) {
-				String base64 = signBase64.substring(start, Math.min(signBase64.length(), start + MAX_LENGTH));
-				Base64Signature base64Signature = DTOUtils.newDTO(Base64Signature.class);
-				base64Signature.setBase64(base64);
-				base64Signature.setApproverInfoId(approverInfo.getId());
-				base64Signature.setOrderNum(base64SignatureList.size());
-				base64SignatureList.add(base64Signature);
-			}
-
-			base64SignatureService.batchInsert(base64SignatureList);
+			this.base64SignatureService.batchInsert(base64SignatureList);
 		}
-		
 
 		return 1;
 	}
 
-	
+	private Optional<ApproverInfo> findByUserName(String userName) {
+		ApproverInfo approverInfo = DTOUtils.newDTO(ApproverInfo.class);
+		approverInfo.setUserName(userName);
+		return this.listByExample(approverInfo).stream().findFirst();
+
+	}
+
+	private List<Base64Signature> buildBase64SignatureList(String signBase64, Long approverInfoId) {
+
+		List<Base64Signature> base64SignatureList = new ArrayList<>();
+
+		for (int start = 0; start < signBase64.length(); start += MAX_LENGTH) {
+			String base64 = signBase64.substring(start, Math.min(signBase64.length(), start + MAX_LENGTH));
+			Base64Signature base64Signature = DTOUtils.newDTO(Base64Signature.class);
+			base64Signature.setBase64(base64);
+			base64Signature.setApproverInfoId(approverInfoId);
+			base64Signature.setOrderNum(base64SignatureList.size());
+			base64SignatureList.add(base64Signature);
+		}
+		return base64SignatureList;
+
+	}
 
 }
